requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gitpython==3.1.40
requests==2.31.0
openai==1.3.7
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
aiofiles==23.2.1
websockets==12.0
docker==6.1.3
pydantic==2.5.1
httpx==0.25.2
GitPython==3.1.40
